refactor(storage): extract apply audit + scheduled runs from sqlite_store
A3 of the sqlite_store decomposition. Two independent extracts:

- amx/storage/_history_apply_audit.py: save_run_results,
  record_evaluation, record_applied, record_db_apply_failure,
  record_apply_event, list_apply_events. Drives the run_results
  write path + apply_events audit log.
- amx/storage/_history_scheduled.py: 11 methods that drive the
  scheduled_runs state machine (create / list / claim_due_schedule /
  arm_next_fire transitions). Tick engine + CLI schedule commands.

Each function takes the store as 'hs' and reuses _lock + _connect().
No DDL touched. SQLiteHistoryStore keeps thin delegating wrappers on
every name so the dozens of callers stay byte-compatible.

sqlite_store.py: 3019 -> 2396 LOC (-623, -21%). Combined with A1+A2,
sqlite_store.py is now 3535 -> 2396 LOC (-32%) over three PRs.
